Time Slot Ease of Use
Updated the Ease of Use for Time Slots in Time Table by taking the End Time of prev Slot as the Start time of Next Slot. Automatically Updates the End time as Start Time + 1 hr by Default. The Default Time Slots to be used are calculated Seperate for each day.